format sign-in page ui
