+++ src/server/db.ts
@@ -1,7 +1,6 @@
 import { PrismaClient as PrismaClientMongo } from "@prisma-db-mongo/client";
 import { PrismaClient as PrismaClientPostgres } from "@prisma-db-psql/client";
 import { type LogLevel } from "@prisma/client/runtime/library";
-import { z } from "zod";
 
 import { env } from "~/env.mjs";
 
@@ -32,12 +31,3 @@ if (NODE_ENV !== "production") {
   globalForPrisma.prismaPostgres = prismaPostgres;
   globalForPrisma.prismaMongo = prismaMongo;
 }
-
-const id_z = z.string().min(1); // can add error message
-const name_z = z.string().min(1);
-const email_z = z.string().email().min(1);
-const emailVerified_z = z.date().nullable();
-const image_z = z.string().nullable();
-const password_z = z.string().min(6);
-
-export { id_z, name_z, email_z, emailVerified_z, image_z, password_z };
